+++ src/generator/models/stargazer-get-request.dto.ts
